+++ test/E2E/TestFunctionApp/DurableOrchestratorGetTaskResult/run.ps1
@@ -1,0 +1,8 @@
+param($Context)
+
+$output = @()
+
+$task = Invoke-DurableActivity -FunctionName 'DurableActivity' -Input "world" -NoWait
+$firstTask = Wait-DurableTask -Task @($task) -Any
+$output += Get-DurableTaskResult -Task @($firstTask)
+$output